requirements.txt
streamlit
google-generativeai
PyPDF2
python-dotenv
scikit-learn
torch
transformers
farm-haystack